requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMatting==1.1.14
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rembg==2.0.67
requests==2.32.4
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.15.3
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
slowapi
slowapi[fastapi]
# ORM y PostgreSQL
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
# Hasheo de contraseñas
passlib[bcrypt]==1.7.4
# JWT
python-jose[cryptography]==3.3.0
pydantic[email]
pydantic-settings